refactor: move account and log viewer components to components directory
