+++ src/components/log-viewer.tsx
